requirements.txt
requests>=2.32.0
python-dotenv>=1.0.1
wordcloud>=1.9.3
matplotlib>=3.9.0
pandas>=2.2.2
jieba>=0.42.1
deep-translator>=1.11.4
streamlit>=1.28.0
pillow>=10.0.0
